LayoutTests/platform/chromium-win/plugins/mouse-events-expected.txt

CONSOLE MESSAGE: line 17: PLUGIN: getFocusEvent
CONSOLE MESSAGE: line 17: PLUGIN: mouseDown at (20, 20)
CONSOLE MESSAGE: line 18: PLUGIN: mouseUp at (20, 20)
CONSOLE MESSAGE: line 20: PLUGIN: mouseDown at (30, 30)
CONSOLE MESSAGE: line 22: PLUGIN: mouseUp at (40, 30)

Test for bug 11517: Flash clicks/interactivity not working properly.
